+++ athena/core/interfaces/__init__.py
@@ -1,5 +1,4 @@
-from athena.core.interfaces.candle import Candle
 from athena.core.interfaces.dataset_layout import DatasetLayout
 from athena.core.interfaces.fluctuations import Fluctuations
 
-__all__ = ["Candle", "Fluctuations", "DatasetLayout"]
+__all__ = ["Fluctuations", "DatasetLayout"]
