+++ tests/core/interfaces/test_fluctuations_functions.py
@@ -3,7 +3,7 @@ import logging
 
 import pytest
 
-from athena.core.interfaces import Candle
+from athena.core.market_entities import Candle
 from athena.core.interfaces.fluctuations import (
     _convert_candles_to_period,
     _merge_candles,
