+++ tests/tradingtools/indicators/conftest.py
@@ -2,7 +2,8 @@ import datetime
 
 import pytest
 
-from athena.core.interfaces import Candle, Fluctuations
+from athena.core.interfaces.fluctuations import Fluctuations
+from athena.core.market_entities import Candle
 from athena.core.types import Coin, Period
